+++ src/features/news/ui/index.tsx
@@ -1,9 +1,4 @@
-'use client';
-
-import Image from 'next/image';
-import { memo } from 'react';
-
-export const NewsImage = memo(({ src }: { src: string }) => (
-  <Image src={src} alt="" priority fill style={{ objectFit: 'cover' }} />
-));
-NewsImage.displayName = 'NewsImage';
+export { default as NewsImage } from './NewsImage';
+export { default as NewsItem } from './NewsItem';
+export { default as NewsItemMeta } from './NewsItemMeta';
+export { default as NewsList } from './NewsList';
